refactor: 에러 코드 파일 분리
