+++ src/errors/CustomError.js
@@ -1,0 +1,8 @@
+class CustomError extends Error {
+  constructor(message, name) {
+    super(message);
+    this.name = name;
+  }
+}
+
+export default CustomError;
